refactor(suite): phishing detection
