+++ suite-common/token-definitions/src/index.ts
@@ -1,9 +1,9 @@
+export * from './phishing';
 export * from './tokenDefinitionsActions';
 export * from './tokenDefinitionsConstants';
-export * from './tokenDefinitionsSelectors';
-export * from './tokenDefinitionsReducer';
-export * from './tokenDefinitionsThunks';
 export * from './tokenDefinitionsMiddleware';
+export * from './tokenDefinitionsReducer';
+export * from './tokenDefinitionsSelectors';
+export * from './tokenDefinitionsThunks';
 export * from './tokenDefinitionsTypes';
 export * from './tokenDefinitionsUtils';
-export * from './antiFraud';
