+++ suite-common/token-definitions/src/phishing/types.ts
@@ -1,0 +1,25 @@
+import type { WalletAccountTransaction } from '@suite-common/wallet-types';
+import { InternalTransfer, TokenTransfer } from '@trezor/blockchain-link-types';
+
+import type { TokenDefinitions } from '../tokenDefinitionsTypes';
+
+export interface TokenTransferWithFiatAmount extends TokenTransfer {
+    amountInFiat?: string;
+}
+
+export interface InternalTransferWithFiatAmount extends InternalTransfer {
+    amountInFiat?: string;
+}
+
+export interface TransactionWithFiatAmount extends WalletAccountTransaction {
+    amountInFiat?: string;
+    tokens: TokenTransferWithFiatAmount[];
+    internalTransfers: InternalTransferWithFiatAmount[];
+}
+
+type PhishingDetectorFnProps = {
+    transaction: TransactionWithFiatAmount;
+    tokenDefinitions?: TokenDefinitions;
+};
+
+export type PhishingDetectorFn = (props: PhishingDetectorFnProps) => boolean;
